% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci.R
\docType{package}
\name{citsci-package}
\alias{citsci}
\alias{citsci-package}
\title{CitSci: An R library for accessing the CitSci.org API}
\description{
\url{https://CitSci.org} is a citizen science platform for collecting scientific observations from the community.
This library is used to access its API using R scripts. This is in very early development & will contain bugs. Additional developer information can be found at \url{https://citsci.org/developers}
The source code for this library may be found at \url{ https://github.com/BBudnicki/citscir }
}
\seealso{
Useful links:
\itemize{
  \item \url{https://citsci.org}
  \item \url{https://citsci.org/developers}
  \item \url{https://github.com/BBudnicki/citscir}
}

}
