% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getObservations.R
\name{getObservations}
\alias{getObservations}
\title{getObservations}
\usage{
getObservations(
  env = "production",
  project = "7a84449a-5b34-41e6-94c9-bfd0570ed7ef"
)
}
\value{
Array of observations
}
\description{
Get observations based on criteria
}
\examples{
getPackageInformation('7a84449a-5b34-41e6-94c9-bfd0570ed7ef')


}
