% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{getPackageInformation}
\alias{getPackageInformation}
\title{getPackageInformation}
\usage{
getPackageInformation()
}
\value{
Information about this package
}
\description{
View metadata about the CitSci R package \url{https://citsci.org/developers}
}
\examples{
getPackageInformation()


}
